requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.5.1
chromadb==0.4.18
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.1

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

